requirements.txt
numpy==1.22.4
pandas==1.5.3
h5py==3.10.0
seaborn==0.12.2
scikit-learn==1.3.2
scikit-misc==0.2.0
annoy
hnswlib
scib==1.1.4
scanpy==1.9.6
squidpy==1.2.3
harmony-pytorch==0.1.7
ipykernel==6.27.0
jupyterlab==4.0.9
rpy2==3.5.14